+++ src/student/student.service.ts
@@ -2,8 +2,9 @@ import { BadRequestException, HttpCode, Injectable, InternalServerErrorException
 import { InjectModel } from '@nestjs/mongoose';
 import { Student, StudentSchema } from './schemas/student.schema';
 import * as mongoose from 'mongoose';
-import { UserMiddleware } from '../middleware/user.middleware';
+import { UserHelper } from '../helper/user.helper';
 import { Attendance } from '../attendance/schemas/attendance.schema';
+import { logger } from '../logger/logger.service';
 
 @Injectable()
 export class StudentService {
@@ -21,11 +22,12 @@ export class StudentService {
      */
     async findAll() : Promise<Student[]>{
         const students = await this.StudentModel.find({})
-        const publicStudent = new UserMiddleware()
+        //const publicStudent = new UserMiddleware()
         if(process.env.NODE_ENV === 'test'){
           return students
         }
-        const secureStudents = students.map(student => publicStudent.getPublicProfile(student))
+        const secureStudents = students.map(student => UserHelper.getPublicProfile(student))
+        logger.info(`successfully print all data of student`)
         return secureStudents;
     }
 
@@ -40,8 +42,8 @@ export class StudentService {
       if(process.env.NODE_ENV === 'test'){
         return student
       }
-      const publicStudent = new UserMiddleware()
-      publicStudent.getPublicProfile(student)
+      UserHelper.getPublicProfile(student)
+      logger.info(`successfully find Student of _id: ${id}`)
       return student;
     }
 
@@ -53,17 +55,18 @@ export class StudentService {
      */
     async createOne(studentData : Student) : Promise<Student> {
       if(process.env.NODE_ENV !== 'test'){
-        const publicStudent = new UserMiddleware()
-        const hashedpasswordStudent = await publicStudent.convertToHash(studentData)
+        const hashedpasswordStudent = await UserHelper.convertToHash(studentData)
         const newStudent = new this.StudentModel(hashedpasswordStudent)
         if(!newStudent){
+          logger.error(`Bad Request Exception`)
           throw new BadRequestException('Enter valid Studentdata ')
         }
         try {
-          const tokenGenerator = new UserMiddleware()
-          tokenGenerator.generateAuthToken(newStudent)
+          UserHelper.generateAuthToken(newStudent)
+          logger.info(`new Student is created of student id : ${newStudent._id}`)
           return newStudent
         } catch(e){
+          logger.error(`Error generate : ${e}`)
           console.log(e)
         }
       } else{
@@ -82,21 +85,24 @@ export class StudentService {
       if(process.env.NODE_ENV !== 'test'){
         const updatable = ['name', 'email', 'currentSem', 'password', 'phoneNumber', 'batch', 'attendance', 'department']
         const updateStudent = Object.keys(studentdata)
+        //check for up update is valid or not
         const isValidUpdate = updateStudent.every(update => updatable.includes(update))
         if(!isValidUpdate){
+          logger.error(` Not valid Update for student`)
           throw new BadRequestException('not valid Update')
         }
-        const publicStudent = new UserMiddleware()
-        if(studentdata.hasOwnProperty('password')){
-          studentdata = await publicStudent.convertToHash(studentdata)
+        // if update is passwaord then convert it to hashpassword
+        if(studentdata.hasOwnProperty('password')){     
+          studentdata = await UserHelper.convertToHash(studentdata)
           console.log(studentdata)
         }
         const updatedStudent = await this.StudentModel.findByIdAndUpdate(id, studentdata)
         if(process.env.NODE_ENV === 'test'){
           return updatedStudent
         }
-        publicStudent.getPublicProfile(updatedStudent)
+        UserHelper.getPublicProfile(updatedStudent)
         if(!updatedStudent){
+          logger.error(`Given student of id : ${id} not found`)
           throw new NotFoundException('Given student not found')
         }
         return updatedStudent
@@ -113,7 +119,12 @@ export class StudentService {
     async deleteOne(id : string) {
       const student = await this.StudentModel.findByIdAndDelete(id)
       if(process.env.NODE_ENV !== 'test'){
-        await this.AttendanceModel.deleteMany({ userId : student._id})
+        if(!student){
+          throw new NotFoundException('Given student not found')
+        }
+        //delete attendance releted to that perticular student
+        await this.AttendanceModel.deleteMany({ userId : student._id}) 
+        logger.info(`Succefully delete student of id : ${id}`)
       }
     }
